requirements.txt
requests==2.32.4
schedule==1.2.0
selenium==4.34.2
pandas==2.3.1
numpy==2.3.1
python-dotenv==1.1.1
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2